move packages to respective layers
